backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pdfplumber==0.9.0
opencv-python==4.7.0.72
pydantic==1.10.7
joblib==1.2.0
pandas==2.2.3
numpy<2.0
scikit-learn==1.6.1
scipy==1.15.1
imbalanced-learn==0.13.0  # replaced imblearn==0.0

pytesseract==0.3.10
Pillow==9.5.0

SQLAlchemy==2.0.5  # replaced SQLAlchemy==2.0.5.post2
psycopg2-binary==2.9.6

python-multipart==0.0.5
sendgrid==6.9.7
